네트워크 라이브러리 종료할 때 순차적으로 잘 종료되는지 알기 위해서 로그 추가. CGSFNet 프로젝트에서 경고 제거

+++ NetworkLayer/CGSFNet/ProactorService.cpp
@@ -65,7 +65,7 @@ void ProactorService::handle_read_stream( const ACE_Asynch_Read_Stream::Result& 
 	}
 	else
 	{
-		if (false == ISession::OnReceive(block.rd_ptr(), block.length(), m_sessionDesc))
+		if (false == ISession::OnReceive(block.rd_ptr(), (unsigned short)block.length(), m_sessionDesc))
 		{
 			block.release();
 			ReserveClose();
